Initial folder setup, sample scene edits
The sample scene shows the manner in which objects in the hierarchy should be organized. For example, a room that is made up of many different components e.g. walls, doors, clutter, etc. should be held inside an empty game object. That game object should have no components itself and it should be at position 0, 0, 0 with a rotation of 0, 0, 0, and it is critically important that its scale be 1, 1, 1.

An empty game object that holds a  room and all the objects that make it up should be named appropriately and then placed inside the appropriate top level container, such as "level."